ai-services/requirements.txt
# Minimal FastAPI setup
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Basic ML/AI (without complex dependencies for now)
numpy==1.25.2
pandas==2.1.3